Move runSessionWith (and all that is needed for it to Test.Internal module.
