+++ wai-extra/wai-extra.cabal
@@ -148,6 +148,7 @@ Library
                      Network.Wai.Request
                      Network.Wai.UrlMap
                      Network.Wai.Test
+                     Network.Wai.Test.Internal
                      Network.Wai.EventSource
                      Network.Wai.EventSource.EventStream
   other-modules:     Network.Wai.Middleware.RequestLogger.Internal
